fix vincenty() error with identical points
Setting lambdaP to 0 resulted in an intial error if the two points were
identical. lambdaP is now initialized with a random number between 1 and 2 to
ensure the while loop begins. Added a check for sinSigma==0 to detect identical
points; if true, break the while loop and return a distance of 0.